refactor: extract CalibreBookManager from ModelData [P0-1d]
